tests: Replace SortKey tests on empty strings with a single space
It seems the value returned by CompareInfo.GetSortKey for an empty
string has changed in recent mono versions: it used to be a byte[5], now
it's just a empty byte array. This matches what is returned by .NET on
Windows, so I guess that change was a compatibility fix.

This made 2 of our unit tests fail, so we now do those tests with a
single space instead of an empty string, which is probably more
significant anyway.

Cherry-picked from git master commit adb0a553a5.